Get rid of MySQL to simplify the storage, use better schema typing (#16)

+++ migrations/2025-01-24-152328_rss_items/up.sql
@@ -1,8 +1,7 @@
 -- Your SQL goes here
 CREATE TABLE `rss_items`(
-	`id` INTEGER NOT NULL,
+	`id` BIGINT NOT NULL,
 	`source` TEXT NOT NULL,
-	`created_at` INTEGER NOT NULL,
+	`created_at` BIGINT NOT NULL,
 	PRIMARY KEY(`id`, `source`)
 );
-
